chore: rearrange encode part
